Basic ETL Pipeline/requirements.txt
requests~=2.32
beautifulsoup4~=4.12
pandas~=2.2
coverage~=7.8
pytest~=7.4
google-auth ~=2.36
google-api-python-client ~=2.152